third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-6-expected.txt

Verifies AST of SCSS with @for control statement.

@for $i from 1 through 3 {
    .box#{$i} {
        color: $color;
        margin: 10px;
    }
}
=== AST === 
rule 0: ".box#{$i}"
    property 0
        name: "color"
            range: {"startLine":2,"startColumn":8,"endLine":2,"endColumn":13}
        value: " $color"
            range: {"startLine":2,"startColumn":14,"endLine":2,"endColumn":21}
        range: {"startLine":2,"startColumn":8,"endLine":2,"endColumn":22}
        disabled: false
    property 1
        name: "margin"
            range: {"startLine":3,"startColumn":8,"endLine":3,"endColumn":14}
        value: " 10px"
            range: {"startLine":3,"startColumn":15,"endLine":3,"endColumn":20}
        range: {"startLine":3,"startColumn":8,"endLine":3,"endColumn":21}
        disabled: false
======
Ranges OK.
